requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-text-splitters
groq
pymongo
faiss-cpu
sentence-transformers
python-multipart